+++ sessionize-base/src/main/kotlin/info/metadude/kotlin/library/sessionize/SessionizeService.kt
@@ -1,13 +1,18 @@
 package info.metadude.kotlin.library.sessionize
 
 import info.metadude.kotlin.library.sessionize.gridtable.models.ConferenceDay
-import retrofit2.Call
+import retrofit2.Response
 import retrofit2.http.GET
+import retrofit2.http.Header
 import retrofit2.http.Path
 
 interface SessionizeService {
 
     @GET("api/v2/{apiKey}/view/gridtable")
-    fun getConferenceDays(@Path("apiKey") apiKey: String): Call<List<ConferenceDay>>
+    suspend fun getConferenceDays(
+        @Header("If-None-Match") eTag: String,
+        @Header("If-Modified-Since") lastModifiedAt: String,
+        @Path("apiKey") apiKey: String
+    ): Response<List<ConferenceDay>>
 
 }